+++ apps/api/views/company_views.py
@@ -56,7 +56,7 @@ def create_company(request):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -102,7 +102,7 @@ def read_company(request, pk):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -160,7 +160,7 @@ def update_company(request, pk):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при запросе ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -206,5 +206,5 @@ def delete_company(request, pk):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при запросе ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)