+++ apps/api/views/connect_views.py
@@ -89,7 +89,7 @@ def create_connect(request):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -134,7 +134,7 @@ def read_connect(request, pk):
         return Response(data={'message': 'Такой connect не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -179,7 +179,7 @@ def read_connect_list(request, pk):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -244,7 +244,7 @@ def update_connect(request, pk):
         return Response(data={'message': 'Такой connect не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -289,5 +289,5 @@ def delete_connect(request, pk):
         return Response(data={'message': 'Такой connect не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при создании ' + e.__str__()
+        message = 'Ошибка при обработке запроса ' + e.__str__()
         return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)